+++ scitos2_common/scitos2_common-extras.cmake
@@ -1,0 +1,18 @@
+# Copyright (c) 2024 Alberto J. Tudela Roldán
+# Copyright (c) 2024 Grupo Avispa, DTE, Universidad de Málaga
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+set(AMENT_BUILD_CONFIGURATION_KEYWORD_SEPARATOR ":")
+
+include("${scitos2_common_DIR}/mira_package.cmake")