+++ scitos2_core/CHANGELOG.rst
@@ -2,10 +2,11 @@
 Changelog for package scitos2_core
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-2.0.0 (25-03-2024)
+2.0.0 (XX-04-2024)
 ------------------
 * Initial release.
 * Create README.md.
 * Added Module class (.hpp file).
 * Added SinkLogger class (.hpp file).
+* Added dependency to scitos2_common.
 * Contributors: Alberto Tudela