+++ scitos2_modules/CHANGELOG.rst
@@ -2,7 +2,8 @@
 Changelog for package scitos2_modules
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-2.0.0 (25-03-2024)
+2.0.0 (XX-04-2024)
 ------------------
 * Import modules from scitos_mira package.
+* Remove CMakeLists.mira and added dependency to scitos2_common.
 * Contributors: Alberto Tudela.